Prevent CLI support code loader from calling module main exports which are not functions
Before this commit, any support code module was considered to be a step definition initializer and expected to expose a function as its main export. Any other value than a function could not be exposed as the main entry of a module.

Consider the following module:

    var MyModule = {
      something: function() {
        // ...
      }
    };

    module.exports = MyModule;

This caused an error due to `MyModule` being blindly called by the CLI support code loader.

From now on, a function exposed as the main entry of a module is still automatically called by the loader. But if the main entry is something different than a function, it is silently ignored.