Add strict mode
The strict was missing compared to the cucumber ruby and jvm
implementations.

The strict mode flag was added to the configuration classes. The
ast_tree_walker interprets the strict mode and if set, sets 
allFeaturesSucceded to false when witnessing a pending or undefined 
step.

The runtime spec was adjusted to run all test for both strict and non-
strict.

The VolatileConfiguration also was updated. Strict mode is enable if
options has a strict field with true.