Removes color when no tty exists and cli params
Previous when cucumber was piped / used in a tty-less environment the
colour codes ended up being printed, this makes the output unreadable,
mostly the "pretty" formatter.

This does some auto detection based on tty and terminal capabilities and
disables colors when they are unable to be printed.

Also included are cli parameters which can be used to force the colors
on / off.

These are:

```sh
--no-color
--color=false
--color=true
--color=always
--color
```