Configure eslint with prettier
This makes ESlint and Prettier work nicer together. All the formatting
rules of Prettier are properly defined in ESlint config (with the help
of the eslint-config-prettier package). This allows for `eslint --fix`
to make both prettier and eslint happy.

As a side note, the eslint plugin of Atom works super well with this.
When saving a file, it is automatically formatted. It reduces the
cognitive load when writing code, you can ignore missing spaces,
new lines and whatnot, it'll be fixed for you automagically.