Use vanilla js for disable timeouts example
Replaces the current ['Disable timeouts' example](https://github.com/cucumber/cucumber-js/blob/master/docs/support_files/timeouts.md) that uses `bluebird` with an example using a vanilla js `Promise`. It clears the timeout to avoid the hang (took me ages to figure that out). It also shows doing operations after the promise resolves (with a warning).

Not sure this is the most simple example possible as I don't work with promises much, but this is what worked for me and gave me an meaningful understanding of how things work inside there. It could just be hindsight, but I think I would have found this code easier to use as a starting point from which I could build. As the project my team is doing doesn't need `bluebird` for anything else, it would also avoid digging into that library.