Attempt to avoid race condition in FakeReportServer
Two listener on "close" per socket, one destroying it, the other one
resolving the promise when stopping the server.

But the one destroying it may be responsible to also resolving the
promise when stopping the server, so as I understand, it could lead to
resolving the same promise twice for a single socket.