Try a bunch of stuff
It looks like the problem occurs when the FakeReportServer starts up -
sometimes although it says it's ready and listening, the server isn't
really available, however long you wait for it.

We'll need to clean up after this commit, go back and make a more
minimal change, but I wanted to check it in for the record.

Here's what I tried:

- added HttpTerminator for managing the server shutdown (in case it
  wasn't shutting down cleanly)
- added waitOn to poll for the server being up after calling `start()`
- keep app in an instance variable
- use the same port for each server instance

In the end, it looks like the problem was caused by letting the http
server choose its own port (by passing in 0 for the port). Sometimes
this just doesn't seem to give us a working server. If we hard-code
a fixed port, at least while we're also using `waitOn`, it seems to
work consistently.

I'm going to revert this commit next and try to pare down these changes to
isolate the minimal change to fix the problem.