Error in CI with HttpStream/FakeReportServer (#1617)
* Add a test which reproduce the error

* Simplify test to reproduce race condition

* Add logging

* Improve logging

* Try a bunch of stuff

It looks like the problem occurs when the FakeReportServer starts up -
sometimes although it says it's ready and listening, the server isn't
really available, however long you wait for it.

We'll need to clean up after this commit, go back and make a more
minimal change, but I wanted to check it in for the record.

Here's what I tried:

- added HttpTerminator for managing the server shutdown (in case it
  wasn't shutting down cleanly)
- added waitOn to poll for the server being up after calling `start()`
- keep app in an instance variable
- use the same port for each server instance

In the end, it looks like the problem was caused by letting the http
server choose its own port (by passing in 0 for the port). Sometimes
this just doesn't seem to give us a working server. If we hard-code
a fixed port, at least while we're also using `waitOn`, it seems to
work consistently.

I'm going to revert this commit next and try to pare down these changes to
isolate the minimal change to fix the problem.

* Revert "Try a bunch of stuff"

This reverts commit e0563002408d776132c6cbeab9ec757285a66cac.

* Use external library to find a free port for FakeHttpServer

This hopefully resolves the intermittend failures we've had in CI.

Ref #1617

* Remove debug logging

* Undo change from logging

* Simplify server shutdown

No need for all that socket management stuff

* We no longer return a port from FakeReportServer.start

* Simpler / more idiomatic use of promises

* Remove unused method

* Remove race condition test

* Revert "Remove unused method"

This reverts commit 03a5d897f21d9cd98c3a7a89d4974f23f3eb4d93.

Turns out it was being used after all!

* Revert "Remove race condition test"

This reverts commit 5c9599ac0c68a27f5ad691d3aae38c7fa39346a6.

I'd like to see this run once on the pull request before we remove it
for good

* Revert "Revert "Remove race condition test""

This reverts commit 59782d40804e5f6091d7dadc4eb0e6fe9c250792.

We don't want these tests in the main branch, and I don't think we need
them anymore now we've fixed the race condition.

Co-authored-by: Matt Wynne <matt@cucumber.io>